requirements.txt
nuscenes-devkit==1.1.*
opencv-python==4.11.*
numpy==1.26.*
matplotlib==3.5.*
plotly==6.0.*
pyquaternion==0.9.*
ultralytics==8.2.*
scipy==1.13.*
tqdm==4.66.*
